+++ src/ska_mid_cbf_mcs/vcc/vcc_component_manager.py
@@ -18,6 +18,7 @@ from __future__ import annotations  # allow forward references in type hints
 import copy
 import json
 import logging
+import time  # # FIXME
 from typing import Callable, List, Optional, Tuple
 
 # tango imports
@@ -99,6 +100,29 @@ class VccComponentManager(CbfComponentManager, CspObsComponentManager):
         :param receptor_id: Receptor ID
         """
         self._receptor_id = receptor_id
+
+    # FIXME
+    @property
+    def frequency_offset_k(self: VccComponentManager) -> int:
+        """
+        Frequency Offset K-value for this receptor
+
+        :return: the frequency offset k-value
+        """
+        return self._frequency_offset_k
+
+    @frequency_offset_k.setter
+    def frequency_offset_k(
+        self: VccComponentManager, frequency_offset_k: int
+    ) -> None:
+        """
+        Set the frequency offset k-value.
+
+        :param frequency_offset_k: Frequency offset k-value
+        """
+        self._frequency_offset_k = frequency_offset_k
+
+    ##
 
     @property
     def frequency_band(self: VccComponentManager) -> int:
@@ -223,6 +247,7 @@ class VccComponentManager(CbfComponentManager, CspObsComponentManager):
 
         # Initialize attribute values
         self._receptor_id = 0
+        self._frequency_offset_k = 0  # FIXME
 
         self._scan_id = 0
         self._config_id = ""
@@ -383,6 +408,8 @@ class VccComponentManager(CbfComponentManager, CspObsComponentManager):
                     for fqdn in self._vcc_band_fqdn
                 ]
 
+                self._init_vcc_controller_parameters()  # FIXME
+
         except tango.DevFailed as df:
             self._logger.error(str(df.args[0].desc))
             self.update_component_fault(True)
@@ -391,6 +418,50 @@ class VccComponentManager(CbfComponentManager, CspObsComponentManager):
         self._logger.info("Completed VccComponentManager.on")
         self.update_component_power_mode(PowerMode.ON)
         return (ResultCode.OK, "On command completed OK")
+
+    # FIXME
+    def _init_vcc_controller_parameters(self: VccComponentManager) -> None:
+        """
+        Initialize the set of parameters in the VCC Controller device that
+        are common to all bands and will not change during scan configuration.
+        """
+        param_init = {
+            "frequency_offset_k": self._frequency_offset_k,
+            "frequency_offset_delta_f": const.DELTA_F,
+        }
+
+        if self._simulation_mode:
+            self._logger.info(
+                "Initializing VCC Controller constant parameters"
+            )
+            self._vcc_controller_simulator.InitCommonParameters(
+                json.dumps(param_init)
+            )
+
+        else:
+            # Wait for VCC Controller
+            for i in range(6):
+                try:
+                    self._vcc_controller_proxy.ping()
+                    break
+                except tango.DevFailed:
+                    time.sleep(5)
+
+            # Skip this if the device has already been initialized
+            if self._vcc_controller_proxy.State() != tango.DevState.INIT:
+                self._logger.info(
+                    "VCC Controller parameters already initialized"
+                )
+                return
+
+            self._logger.info(
+                f"Initializing VCC Controller constant parameters: {param_init}"
+            )
+            self._vcc_controller_proxy.InitCommonParameters(
+                json.dumps(param_init)
+            )
+
+    ##
 
     def off(self: VccComponentManager) -> Tuple[ResultCode, str]:
         """
@@ -417,7 +488,7 @@ class VccComponentManager(CbfComponentManager, CspObsComponentManager):
         return (ResultCode.OK, "Standby command completed OK")
 
     def configure_band(
-        self: VccComponentManager, argin: str
+        self: VccComponentManager, freq_band_name: str  # FIXME
     ) -> Tuple[ResultCode, str]:
         """
         Configure the corresponding band. At the HPS level, this reconfigures the
@@ -434,8 +505,9 @@ class VccComponentManager(CbfComponentManager, CspObsComponentManager):
         (result_code, msg) = (ResultCode.OK, "ConfigureBand completed OK.")
 
         try:
-            band_config = json.loads(argin)
-            freq_band_name = band_config["frequency_band"]
+            # FIXME
+            # band_config = json.loads(argin)
+            # freq_band_name = band_config["frequency_band"]
 
             # Configure the band via the VCC Controller device
             self._logger.info(f"Configuring VCC band {freq_band_name}")
@@ -469,12 +541,13 @@ class VccComponentManager(CbfComponentManager, CspObsComponentManager):
 
             self._logger.info(f"VCC internal parameters: {json_string}")
 
-            args = json.loads(json_string)
-            args.update({"dish_sample_rate": band_config["dish_sample_rate"]})
-            args.update(
-                {"samples_per_frame": band_config["samples_per_frame"]}
-            )
-            json_string = json.dumps(args)
+            # FIXME
+            # args = json.loads(json_string)
+            # args.update({"dish_sample_rate": band_config["dish_sample_rate"]})
+            # args.update(
+            #     {"samples_per_frame": band_config["samples_per_frame"]}
+            # )
+            # json_string = json.dumps(args)
 
             idx = self._freq_band_index[self._freq_band_name]
             if self._simulation_mode:
@@ -522,7 +595,9 @@ class VccComponentManager(CbfComponentManager, CspObsComponentManager):
                 self._vcc_controller_simulator.Unconfigure()
             else:
                 try:
-                    self._vcc_controller_proxy.Unconfigure()
+                    pass
+                    # TODO CIP-1850
+                    # self._vcc_controller_proxy.Unconfigure()
                 except tango.DevFailed as df:
                     self._logger.error(str(df.args[0].desc))
                     self.update_component_fault(True)
@@ -661,7 +736,9 @@ class VccComponentManager(CbfComponentManager, CspObsComponentManager):
                 self._band_simulators[idx].Abort()
             else:
                 try:
-                    self._band_proxies[idx].Abort()
+                    pass
+                    # TODO CIP-1850
+                    # self._band_proxies[idx].Abort()
                 except tango.DevFailed as df:
                     self._logger.error(str(df.args[0].desc))
                     self.update_component_fault(True)
@@ -686,7 +763,9 @@ class VccComponentManager(CbfComponentManager, CspObsComponentManager):
                 self._band_simulators[idx].ObsReset()
             else:
                 try:
-                    self._band_proxies[idx].ObsReset()
+                    pass
+                    # TODO CIP-1850
+                    # self._band_proxies[idx].ObsReset()
                 except tango.DevFailed as df:
                     self._logger.error(str(df.args[0].desc))
                     self.update_component_fault(True)