+++ tests/unit/controller_component_manager/ControllerComponentManager_test.py
@@ -87,11 +87,10 @@ class TestControllerComponentManager:
         (result_code, _) = controller_component_manager.init_sys_param(sp)
         assert result_code == ResultCode.OK
 
-        # TODO: CIP-1569
-        # Temporarily commenting out the assert below
-        # assert controller_component_manager._on is False
+        assert controller_component_manager._on is False
 
         (result_code, _) = controller_component_manager.on()
+        assert controller_component_manager._on is True
         assert result_code == ResultCode.OK
 
     def test_On_No_SysParam(
@@ -129,11 +128,8 @@ class TestControllerComponentManager:
         )
         assert controller_component_manager._connected is True
 
-        # TODO: CIP-1569
-        # Temporarily commenting out the next two lines
-        # (result_code, _) = controller_component_manager.on()
-        #
-        # assert controller_component_manager._on is True
+        (result_code, _) = controller_component_manager.on()
+        assert controller_component_manager._on is True
 
         (result_code, _) = controller_component_manager.off()
         assert result_code == ResultCode.OK